wip: restructure, rework config
