+++ apps/nspanel_haui/haui/abstract/entity.py
@@ -1,24 +1,21 @@
 from copy import deepcopy
-from typing import Dict, List
 
-from .mapping.color import COLORS
-from .mapping.const import (
-    DEFAULT_CONFIG,
-    PANEL_CONFIG,
+from ..mapping.color import COLORS
+from ..mapping.const import (
     ENTITY_CONFIG,
     INTERNAL_ENTITY_TYPE,
 )
 
-from .helper.color import rgb_to_rgb565
-from .helper.entity import (
+from ..helper.color import rgb_to_rgb565
+from ..helper.entity import (
     get_entity_icon,
     get_entity_color,
     get_entity_name,
     get_entity_value,
     execute_entity,
 )
-from .helper.text import get_state_translation
-from .helper.value import merge_dicts
+from ..helper.text import get_state_translation
+from ..helper.value import merge_dicts
 
 from .base import HAUIBase
 
@@ -356,248 +353,3 @@ class HAUIConfigEntity(HAUIBase):
         return get_state_translation(
             self.get_entity_type(), self.get_entity_state(), self.get_locale()
         )
-
-
-class HAUIConfigPanel(HAUIBase):
-
-    """
-    Represents a panel on the display.
-
-    Its a description of what the page should look like and what
-    entities to show
-    """
-
-    def __init__(self, app, config=None):
-        """Initialize for config panel.
-
-        Args:
-            app (NSPanelHAUI): App
-            config (dict, optional): Config for panel. Defaults to None.
-        """
-        # initialize with default panel values
-        super().__init__(app, deepcopy(PANEL_CONFIG))
-        # load config
-        if config is not None:
-            merge_dicts(self._config, config)
-        # store the initial config in an additional var
-        # so it is possible to restore the config to initial values
-        self._default_config = {}
-        merge_dicts(self._default_config, self._config)
-        # store persistent config
-        self._persistent_config = {}
-        # load all entities
-        self._entities: List[HAUIConfigEntity] = []  # list of HAUIConfigEntity
-        # single entity config
-        if self.get("entity") is not None:
-            self._entities.append(HAUIConfigEntity(self.app, config))
-        # multiple entities config
-        for e in self.get("entities", []):
-            self._entities.append(HAUIConfigEntity(self.app, e))
-
-    # public
-
-    def get_type(self) -> str:
-        """Returns the type of this panel.
-
-        Returns:
-            str: Type
-        """
-        return self.get("type", "")
-
-    def get_mode(self) -> str:
-        """Returns the panel mode.
-
-        Possible panel modes:
-        - panel (Default)
-        - subpanel
-        - popup
-
-        Returns:
-            str: Panel mode
-        """
-        return self.get("mode", "panel")
-
-    def get_title(self, default_title=None) -> str:
-        """Returns the title of this panel.
-
-        Args:
-            default_title (str, optional): Default title if not set. Defaults to None.
-
-        Returns:
-            str: Title
-        """
-        if default_title is None:
-            default_title = self.translate("Unnamed")
-        title = self.get("title", "")
-        if title == "":
-            title = default_title
-        return title
-
-    def is_home_panel(self) -> bool:
-        """Returns True if this is the home panel.
-
-        Returns:
-            bool: True if home panel
-        """
-        return self.get("home_panel", False)
-
-    def is_sleep_panel(self) -> bool:
-        """Returns True if this is the sleep panel.
-
-        Returns:
-            bool: True if sleep panel
-        """
-        return self.get("sleep_panel", False)
-
-    def is_wakeup_panel(self) -> bool:
-        """Returns True if this is the wakeup panel.
-
-        Returns:
-            bool: True if wakeup panel
-        """
-        return self.get("wakeup_panel", False)
-
-    def show_home_button(self) -> bool:
-        """Returns True if home button should be shown.
-
-        Returns:
-            bool: True if home button should be shown
-        """
-        show_home_button = self.get(
-            "show_home_button", self.app.device.get("show_home_button", False)
-        )
-        return show_home_button
-
-    def get_entities(self, return_copy=True) -> List[HAUIConfigEntity]:
-        """Returns all entities from this panel.
-
-        Args:
-            return_copy (bool, optional): Copy entities. Defaults to True.
-
-        Returns:
-            list: List with entities
-        """
-        if return_copy:
-            return self._entities.copy()
-        return self._entities
-
-    def get_default_config(self, return_copy=True) -> Dict:
-        """Returns the initial config of this panel.
-
-        Returns:
-            dict: Initial config
-            return_copy (bool, optional): Return copy of config. Defaults to True.
-        """
-        if return_copy:
-            return self._default_config.copy()
-        return self._default_config
-
-    def get_persistent_config(self, return_copy=True) -> Dict:
-        """Returns the persistent config of this panel.
-
-        Returns:
-            dict: Persistent config
-            return_copy (bool, optional): Return copy of config. Defaults to True.
-        """
-        if return_copy:
-            return self._persistent_config.copy()
-        return self._persistent_config
-
-    def restore_default_config(self) -> Dict:
-        """Restore the initial config of this panel."""
-        self._config = self.get_default_config(return_copy=True)
-
-
-class HAUIConfig(HAUIBase):
-
-    """
-    Configuration helper
-    """
-
-    def __init__(self, app, config=None):
-        """Initialize for config.
-
-        Args:
-            app (NSPanelHAUI): App
-            config (dict, optional): Config. Defaults to None.
-        """
-        # initialize with default config
-        super().__init__(app, deepcopy(DEFAULT_CONFIG))
-        # load config
-        if config is not None:
-            merge_dicts(self.get_config(return_copy=False), config)
-        # load all panels
-        self._panels: List[HAUIConfigPanel] = []
-        panels_to_load = self.get("panels", [])
-        # append sys_panels so they can be overwritten by panels
-        panels_to_load += self.get("sys_panels", [])
-        for panel_config in panels_to_load:
-            panel = HAUIConfigPanel(self.app, panel_config)
-            self._panels.append(panel)
-
-    # public
-
-    def get_panels(self, filter_nav_panel=None) -> List[HAUIConfigPanel]:
-        """Returns all panels as HAUIConfigPanel objects.
-
-        Args:
-            filter_nav_panel (bool, optional): Filter panels panel_nav attr. Defaults to None.
-
-        Returns:
-            list: List with panels
-        """
-        if filter_nav_panel is not None:
-            # provide a filtered list if nav_panel provided
-            # True means only nav_panels will be returned, False non nav_panels
-            nav_panels = list(
-                filter(
-                    lambda panel: (filter_nav_panel and panel.get_mode() == "panel")
-                    or (not filter_nav_panel and panel.get_mode() != "panel"),
-                    self._panels,
-                )
-            )
-            return nav_panels
-        return self._panels
-
-    def get_entities(self) -> List[HAUIConfigEntity]:
-        """Returns all entities as HAUIConfigEntity objects.
-
-        Returns:
-            list: List with entities
-        """
-        entities = []
-        for panel in self._panels:
-            entities.extend(panel.get_entities())
-        return entities
-
-    def get_entity(self, entity_id) -> HAUIConfigEntity:
-        """Returns a single entity.
-
-        Args:
-            entity_id (str): Entity id
-
-        Returns:
-            HAUIConfigEntity: Entity
-        """
-        for entity in self.get_entities():
-            if entity.id == entity_id:
-                return entity
-        return None
-
-    def get_panel(self, panel_id) -> HAUIConfigPanel:
-        """Returns a single panel.
-
-        Args:
-            panel_id (str): Panel id or key
-
-        Returns:
-            HAUIConfigPanel: Panel
-        """
-        for panel in self._panels:
-            # get by id
-            if panel.id == panel_id:
-                return panel
-            # get by key
-            if panel.get("key", "") == panel_id:
-                return panel
-        return None
