+++ apps/nspanel_haui/haui/controller/mqtt.py
@@ -1,7 +1,8 @@
 import json
 
 from ..mapping.const import ALL_RECV, ALL_CMD
-from ..base import HAUIPart, HAUIEvent
+from ..abstract.part import HAUIPart
+from ..abstract.event import HAUIEvent
 
 
 class HAUIMQTTController(HAUIPart):
