+++ apps/nspanel_haui/haui/page/about.py
@@ -1,6 +1,7 @@
 import haui.version
 
-from ..config import HAUIConfigPanel
+from ..abstract.panel import HAUIConfigPanel
+
 from . import HAUIPage
