+++ apps/nspanel_haui/haui/page/blank.py
@@ -1,4 +1,5 @@
-from ..config import HAUIConfigPanel
+from ..abstract.panel import HAUIConfigPanel
+
 from . import HAUIPage
