+++ apps/nspanel_haui/haui/page/notify.py
@@ -1,6 +1,6 @@
 from ..mapping.color import COLORS
 from ..helper.icon import parse_icon
-from ..config import HAUIConfigPanel
+from ..abstract.panel import HAUIConfigPanel
 
 from . import HAUIPage
