+++ apps/nspanel_haui/haui/page/qr.py
@@ -1,5 +1,5 @@
 from ..mapping.color import COLORS
-from ..config import HAUIConfigPanel
+from ..abstract.panel import HAUIConfigPanel
 
 from . import HAUIPage
